# improve performance for repeatable records