- Improve performance of pipe-all button for repeatable records. These are records of projects that have:
  - Repeatable events
  - Repeatable instruments

**Note**: There's no performance improvement for non-repeatable records yet for now.

- Improvement details:
  - Remove a loop when checking if record is repeatable (details [here](https://github.com/vanderbilt-redcap/cross-project-piping-module/commit/ffdae9ec0b73d82e154217a404645c6a7d846cbc#diff-ee0a829d9982fc7786b79bdca83b0ac441e0145dac8df47c9803a074f8cf0f56R977-R981)).
  - For each destination piping field, check if its value already equals the source field's. If it does don't pipe it (details [here](https://github.com/vanderbilt-redcap/cross-project-piping-module/commit/ffdae9ec0b73d82e154217a404645c6a7d846cbc#diff-ee0a829d9982fc7786b79bdca83b0ac441e0145dac8df47c9803a074f8cf0f56R1036-R1044)).

- Benchmark (~1500 records):

|              Performance | Non-repeatable | Repeatable instruments | Repeatable events |
|-------------------------:|----------------|------------------------|-------------------|
|  Original full data avg. | 165s           |          297s          |        180s       |
| Improved no updates avg. | N/A            |          84.5s         |        27s        |
|   Improved 1 update avg. | N/A            |           88s          |       27.5s       |
|  Improved full data avg. | 164s           |          220s          |        86s        |
|     Full data difference | +1s            |          -77s          |        -94s       |